Created Router and separate every request to different files
